148 Add BranchingBackend to support running scripts
