+++ netbox_branching/utilities.py
@@ -4,12 +4,17 @@ from collections import defaultdict
 from contextlib import contextmanager
 from dataclasses import dataclass
 
+from django.contrib import messages
+from django.core.exceptions import ObjectDoesNotExist
 from django.db.models import ForeignKey, ManyToManyField
+from django.http import HttpResponseBadRequest
 from django.urls import reverse
 
 from netbox.plugins import get_plugin_config
 from netbox.registry import registry
+from .choices import BranchStatusChoices
 from .constants import EXEMPT_MODELS, INCLUDE_MODELS
+from .constants import COOKIE_NAME, BRANCH_HEADER, QUERY_PARAM
 from .contextvars import active_branch
 
 __all__ = (
@@ -209,4 +214,39 @@ def is_api_request(request):
     """
     Returns True if the given request is a REST or GraphQL API request.
     """
+    if not hasattr(request, 'path_info'):
+        return False
+
     return request.path_info.startswith(reverse('api-root')) or request.path_info.startswith(reverse('graphql'))
+
+
+def get_active_branch(request):
+    """
+    Return the active Branch (if any).
+    """
+    # The active Branch may be specified by HTTP header for REST & GraphQL API requests.
+    from .models import Branch
+    if is_api_request(request) and BRANCH_HEADER in request.headers:
+        branch = Branch.objects.get(schema_id=request.headers.get(BRANCH_HEADER))
+        if not branch.ready:
+            return HttpResponseBadRequest(f"Branch {branch} is not ready for use (status: {branch.status})")
+        return branch
+
+    # Branch activated/deactivated by URL query parameter
+    elif QUERY_PARAM in request.GET:
+        if schema_id := request.GET.get(QUERY_PARAM):
+            branch = Branch.objects.get(schema_id=schema_id)
+            if branch.ready:
+                messages.success(request, f"Activated branch {branch}")
+                return branch
+            else:
+                messages.error(request, f"Branch {branch} is not ready for use (status: {branch.status})")
+                return None
+        else:
+            messages.success(request, f"Deactivated branch")
+            request.COOKIES.pop(COOKIE_NAME, None)  # Delete cookie if set
+            return None
+
+    # Branch set by cookie
+    elif schema_id := request.COOKIES.get(COOKIE_NAME):
+        return Branch.objects.filter(schema_id=schema_id, status=BranchStatusChoices.READY).first()
